LayoutTests/webgl/2.0.0/conformance2/transform_feedback/transform_feedback-expected.txt

This test runs the WebGL Test listed below in an iframe and reports PASS or FAIL.

Test: ../../resources/webgl_test_files/conformance2/transform_feedback/transform_feedback.html
[ 1: PASS ] WebGL context exists
[ 2: PASS ] gl.TRANSFORM_FEEDBACK_BINDING is 0x8E25
[ 3: PASS ] getError was expected value: NO_ERROR : TRANSFORM_FEEDBACK_BINDING query should succeed
[ 4: PASS ] gl.getParameter(gl.TRANSFORM_FEEDBACK_BINDING) is null
[ 5: PASS ] gl.getParameter(gl.TRANSFORM_FEEDBACK_BINDING) is null
[ 6: PASS ] gl.getParameter(gl.TRANSFORM_FEEDBACK_BINDING) is tf
[ 7: PASS ] gl.getParameter(gl.TRANSFORM_FEEDBACK_BINDING) is tf1
[ 8: PASS ] gl.getParameter(gl.TRANSFORM_FEEDBACK_BINDING) is null
[ 9: PASS ] getError was expected value: INVALID_OPERATION : binding a deleted Transform Feedback object
[ 10: PASS ] gl.getParameter(gl.TRANSFORM_FEEDBACK_BINDING) is null
[ 11: PASS ] getError was expected value: NO_ERROR : there should be no errors
[ 12: FAIL ] gl.getIndexedParameter(gl.TRANSFORM_FEEDBACK_BUFFER_BINDING, 0) should be null. Was [object WebGLBuffer].
[ 13: PASS ] getError was expected value: NO_ERROR : binding buffer to ARRAY_BUFFER
[ 14: PASS ] getError was expected value: NO_ERROR : createTransformFeedback should not set an error
[ 15: PASS ] tf is non-null.
[ 16: PASS ] gl.isTransformFeedback(tf) is false
[ 17: PASS ] gl.isTransformFeedback(tf) is true
[ 18: PASS ] gl.isTransformFeedback(tf) is true
[ 19: PASS ] gl.isTransformFeedback(tf) is false
[ 20: PASS ] gl.isTransformFeedback(null) is false
[ 21: PASS ] getError was expected value: NO_ERROR : linking transform feedback shader should not set an error
[ 22: PASS ] program is non-null.
[ 23: PASS ] getError was expected value: NO_ERROR : after evaluating: gl.getBufferSubData(gl.TRANSFORM_FEEDBACK_BUFFER, 0, retArray, 0, retArray.length)
[ 24: PASS ] getError was expected value: NO_ERROR : after evaluating: gl.getBufferSubData(gl.TRANSFORM_FEEDBACK_BUFFER, 0, retArray, 0, retArray.length)
[ 25: PASS ] getError was expected value: INVALID_OPERATION : after evaluating: gl.getBufferSubData(gl.TRANSFORM_FEEDBACK_BUFFER, 0, retArray, 0, retArray.length)
[ 26: PASS ] buffer should match expected values
[ 27: PASS ] getError was expected value: NO_ERROR : linking transform feedback shader should not set an error
[ 28: PASS ] program is non-null.
[ 29: PASS ] buffer should match expected values
[ 30: PASS ] getError was expected value: NO_ERROR : linking transform feedback shader should not set an error
[ 31: PASS ] program is non-null.
[ 32: PASS ] buffer should match expected values
[ 33: PASS ] buffer should match expected values
[ 34: PASS ] TRANSFORM_FEEDBACK_PRIMITIVES_WRITTEN query's result didn't become available too early
[ 35: PASS ] TRANSFORM_FEEDBACK_PRIMITIVES_WRITTEN query returned a correct result (3)
[ 36: PASS ] getError was expected value: NO_ERROR : linking transform feedback shader should not set an error
[ 37: PASS ] program is non-null.
[ 38: PASS ] gl.getProgramParameter(program, gl.TRANSFORM_FEEDBACK_VARYINGS) is 2
[ 39: PASS ] getError was expected value: NO_ERROR : Should be no errors from valid getTransformFeedbackVarying.
[ 40: PASS ] activeInfo is non-null.
[ 41: PASS ] activeInfo.name is 'out_add'
[ 42: PASS ] activeInfo.type is gl.FLOAT_VEC4
[ 43: PASS ] activeInfo.size is 1
[ 44: PASS ] getError was expected value: NO_ERROR : Should be no errors from valid getTransformFeedbackVarying.
[ 45: PASS ] activeInfo is non-null.
[ 46: PASS ] activeInfo.name is 'out_mul'
[ 47: PASS ] activeInfo.type is gl.FLOAT_VEC4
[ 48: PASS ] activeInfo.size is 1
[ 49: PASS ] getError was expected value: INVALID_VALUE : Should be INVALID_VALUE when calling getTransformFeedbackVarying with an invalid index.
[ 50: PASS ] activeInfo is null
[ 51: PASS ] gl.getProgramParameter(program, gl.TRANSFORM_FEEDBACK_VARYINGS) is 2
[ 52: PASS ] getError was expected value: NO_ERROR : Should be no errors from valid getTransformFeedbackVarying.
[ 53: PASS ] activeInfo is non-null.
[ 54: PASS ] activeInfo.name is 'out_add'
[ 55: PASS ] activeInfo.type is gl.FLOAT_VEC4
[ 56: PASS ] activeInfo.size is 1
[ 57: PASS ] getError was expected value: NO_ERROR : Should be no errors from valid getTransformFeedbackVarying.
[ 58: PASS ] activeInfo is non-null.
[ 59: PASS ] activeInfo.name is 'out_mul'
[ 60: PASS ] activeInfo.type is gl.FLOAT_VEC4
[ 61: PASS ] activeInfo.size is 1
[ 62: PASS ] getError was expected value: INVALID_VALUE : Should be INVALID_VALUE when calling getTransformFeedbackVarying with an invalid index.
[ 63: PASS ] activeInfo is null
[ 64: PASS ] getError was expected value: NO_ERROR : linking transform feedback shader should not set an error
[ 65: PASS ] gl.getProgramParameter(program, gl.LINK_STATUS) is true
[ 66: PASS ] gl.getProgramParameter(program, gl.TRANSFORM_FEEDBACK_VARYINGS) is 1
[ 67: PASS ] getError was expected value: NO_ERROR : Should be no errors from valid getTransformFeedbackVarying.
[ 68: PASS ] activeInfo is non-null.
[ 69: PASS ] activeInfo.name is 'out_mul'
[ 70: PASS ] activeInfo.type is gl.FLOAT_VEC4
[ 71: PASS ] activeInfo.size is 1
[ 72: PASS ] getError was expected value: INVALID_VALUE : Should be INVALID_VALUE when calling getTransformFeedbackVarying with an invalid index.
[ 73: PASS ] activeInfo is null
[ 74: PASS ] getError was expected value: INVALID_VALUE : Should be INVALID_VALUE when calling getTransformFeedbackVarying with an invalid index.
[ 75: PASS ] activeInfo is null
[ 76: PASS ] getError was expected value: NO_ERROR : linking transform feedback shader should not set an error
[ 77: PASS ] gl.getProgramParameter(program, gl.LINK_STATUS) is true
[ 78: PASS ] gl.getProgramParameter(program, gl.TRANSFORM_FEEDBACK_VARYINGS) is 2
[ 79: PASS ] getError was expected value: NO_ERROR : Should be no errors from valid getTransformFeedbackVarying.
[ 80: PASS ] activeInfo is non-null.
[ 81: PASS ] activeInfo.name is 'out_add'
[ 82: PASS ] activeInfo.type is gl.FLOAT_VEC4
[ 83: PASS ] activeInfo.size is 1
[ 84: PASS ] getError was expected value: NO_ERROR : Should be no errors from valid getTransformFeedbackVarying.
[ 85: PASS ] activeInfo is non-null.
[ 86: PASS ] activeInfo.name is 'out_mul'
[ 87: PASS ] activeInfo.type is gl.FLOAT_VEC4
[ 88: PASS ] activeInfo.size is 1
[ 89: PASS ] getError was expected value: INVALID_VALUE : Should be INVALID_VALUE when calling getTransformFeedbackVarying with an invalid index.
[ 90: PASS ] activeInfo is null
[ 91: PASS ] successfullyParsed is true
[ FAIL ] 1 failures reported
